Prevented underflow when computing full conditionals
Just extracted the log of the largest log_prob and
exponentiated, and also set things to 0.0 if they
were going to underflow.  Added a test data set with
lots of SNPs.  The z-specs in that data set don't seem
to be right, but at least it shows that we don't get
the same underflow problems.